Move tree functions to separate file + fixes
